# First Step in optimizing GSP Performance on Grails 7. Safe small optimizations to start