--- a/grails-gsp/core/src/main/groovy/org/grails/gsp/GroovyPageMetaInfo.java
+++ b/grails-gsp/core/src/main/groovy/org/grails/gsp/GroovyPageMetaInfo.java
@@ -22,7 +22,10 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.Writer;
+import java.lang.ref.SoftReference;
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Modifier;
 import java.net.URL;
 import java.net.URLConnection;
@@ -67,14 +70,17 @@ public class GroovyPageMetaInfo implements GrailsApplicationAware {
     private static final Log LOG = LogFactory.getLog(GroovyPageMetaInfo.class);
     private TagLibraryLookup tagLibraryLookup;
     private TagLibraryResolver jspTagLibraryResolver;
+    private ThreadLocal<SoftReference<GroovyPage>> pageInstance = new ThreadLocal<>();
 
     private boolean precompiledMode = false;
     private Class<?> pageClass;
+    private Constructor<?> pageClassConstructor;
     private long lastModified;
     private InputStream groovySource;
     private String contentType;
     private int[] lineNumbers;
     private String[] htmlParts;
+    private Set<Integer> htmlPartsSet;
     @SuppressWarnings("rawtypes")
     private Map jspTags = Collections.emptyMap();
     private GroovyPagesException compilationException;
@@ -115,6 +121,11 @@ public class GroovyPageMetaInfo implements GrailsApplicationAware {
         this();
         precompiledMode = true;
         this.pageClass = pageClass;
+        try {
+            this.pageClassConstructor = pageClass.getConstructor();
+        } catch (NoSuchMethodException e) {
+            throw new RuntimeException(e);
+        }
         contentType = (String) ReflectionUtils.getField(ReflectionUtils.findField(pageClass, GroovyPageParser.CONSTANT_NAME_CONTENT_TYPE), null);
         jspTags = (Map) ReflectionUtils.getField(ReflectionUtils.findField(pageClass, GroovyPageParser.CONSTANT_NAME_JSP_TAGS), null);
         lastModified = (Long) ReflectionUtils.getField(ReflectionUtils.findField(pageClass, GroovyPageParser.CONSTANT_NAME_LAST_MODIFIED), null);
@@ -282,8 +293,26 @@ public class GroovyPageMetaInfo implements GrailsApplicationAware {
         return pageClass;
     }
 
+    public GroovyPage getPageClassInstance() throws InstantiationException, IllegalAccessException, NoSuchMethodException, InvocationTargetException {
+        SoftReference<GroovyPage> pageSoftRef = pageInstance.get();
+
+        GroovyPage pageCacheEntry = pageSoftRef != null ? pageSoftRef.get() : null;
+        if(pageCacheEntry == null) {
+            LOG.info("Loading Page: " + pageClass.getName());
+            pageCacheEntry = (GroovyPage) pageClassConstructor.newInstance();
+            pageCacheEntry.initCommonRun(this);
+            pageInstance.set(new SoftReference<>(pageCacheEntry));
+        }
+        return pageCacheEntry;
+    }
+
     public void setPageClass(Class<?> pageClass) {
         this.pageClass = pageClass;
+        try {
+            this.pageClassConstructor = pageClass.getConstructor();
+        } catch (NoSuchMethodException e) {
+            throw new RuntimeException(e);
+        }
         initializePluginPath();
     }
 
@@ -359,6 +388,18 @@ public class GroovyPageMetaInfo implements GrailsApplicationAware {
 
     public void setHtmlParts(String[] htmlParts) {
         this.htmlParts = htmlParts;
+        this.htmlPartsSet = new HashSet<>();
+        if (htmlParts != null) {
+            for (String htmlPart : htmlParts) {
+                if (htmlPart != null) {
+                    htmlPartsSet.add(System.identityHashCode(htmlPart));
+                }
+            }
+        }
+    }
+
+    Set<Integer> getHtmlPartsSet() {
+        return this.htmlPartsSet;
     }
 
     public void applyLastModifiedFromResource(Resource resource) {