… grails-gsp/plugin/src/test/groovy/org/grails/gsp/compiler/tags/GroovyEachParseTests.groovy
… grails-gsp/plugin/src/test/groovy/org/grails/web/pages/ParseSpec.groovy
… grails-testing-support-web/src/main/groovy/grails/testing/web/GrailsWebUnitTest.groovy