@codeconsole 2025-10-04
Isn't gsp currently exclusive commons logging? 

Not sure if adding org.slf4j dependencies is a good idea if that is the case. 

Spring Framework is 100% [commons logging](https://repo.maven.apache.org/maven2/org/springframework/spring-core/7.0.0-M9/spring-core-7.0.0-M9.pom)