@davydotcom 2025-10-05
This isn't made for spring . It's made for grails