@jdaugherty 2025-10-07
I've reformatted some of the files using IntelliJ to fix the codeStyle errors.  It seems some of this reformat caused parts of the file that were unchanged to be formatted.  This is probably ok, but trying to give a heads up for other users.