@jdaugherty 2025-10-07
> > This isn't made for spring . It's made for grails
> 
> @davydotcom GSP should be independent of Spring and Grails. What is being gained by switching out the logger and requiring another dependency in the classpath?

According to Spring, slf4j is a supported logging platform:  https://docs.spring.io/spring-framework/reference/core/spring-jcl.html