@sbglasius 2025-10-08
I would like to add, that this all looks good and very well thought through.