@davydotcom 2025-10-08
Worth mentioning, this does not break existing GSP compiles. we left the old method names such that the old compiled GSP Files would still function.